+++ packages/features/profile/lib/src/add_new_measurement_bottom_sheet/cubit/add_measurement_cubit.dart
@@ -25,7 +25,7 @@ class AddMeasurementCubit extends Cubit<AddMeasurementState> {
     emit(
       state.copyWith(
         bodyCompositionValues: state.bodyCompositionValues.copyWith(
-          weight: () => int.parse(value),
+          weight: () => double.parse(value),
         ),
       ),
     );
